optimise: add v00 which defines a custom gcd and rename other versions
