+++ optimise/v00.c
@@ -1,6 +1,17 @@
 #include <stdio.h>
-#include <numeric>
 #include <cmath>
+
+
+int euclidean_gcd(int a, int b){
+    int t;
+    while (b != 0){
+        t = b;
+        b = a % b;
+        a = t;
+    }
+    return a;
+}
+
 
 int main(){
     int N, combinations;
@@ -9,7 +20,7 @@ int main(){
         for(int x=1; x < N; x++){
             for(int y=1; y < N; y++){
                 for(int z=1; z < N; z++){
-                    if(std::gcd(std::gcd(x, y), z) == 1 &&
+                    if(euclidean_gcd(euclidean_gcd(x, y), z) == 1 &&
                        std::pow(x, 2) + std::pow(y, 2) == std::pow(z, 2) &&
                        x < y && y < z){
                             combinations += 1;
