+++ optimise/v01.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 #include <numeric>
-
+#include <cmath>
 
 int main(){
     int N, combinations;
@@ -10,7 +10,7 @@ int main(){
             for(int y=1; y < N; y++){
                 for(int z=1; z < N; z++){
                     if(std::gcd(std::gcd(x, y), z) == 1 &&
-                       x * x + y * y == z * z &&
+                       std::pow(x, 2) + std::pow(y, 2) == std::pow(z, 2) &&
                        x < y && y < z){
                             combinations += 1;
                     }
