+++ optimise/v01.py
@@ -9,6 +9,6 @@ for line in sys.stdin:
         for y in range(N + 1):
             for z in range(N + 1):
                 if gcd(gcd(x, y), z) == 1 and\
-                   x * x + y * y == z * z and x < y < z:
+                   x ** 2 + y ** 2 == z ** 2 and x < y < z:
                     combinations += 1
     print(combinations)
