+++ optimise/v02.py
@@ -8,7 +8,7 @@ for line in sys.stdin:
     for x in range(N + 1):
         for y in range(N + 1):
             for z in range(N + 1):
-                if x < y < z and x * x + y * y == z * z and\
-                   gcd(gcd(x, y), z) == 1:
-                        combinations += 1
+                if gcd(gcd(x, y), z) == 1 and\
+                   x * x + y * y == z * z and x < y < z:
+                    combinations += 1
     print(combinations)
