+++ optimise/v03.c
@@ -6,9 +6,9 @@ int main(){
     int N, combinations;
     while (scanf("%d", &N) && N){
         combinations = 0;
-        for(int x=2; x < N; x++){
-            for(int y=x+1; y < N; y++){
-                for(int z=y+1; z < N; z++){
+        for(int x=1; x < N; x++){
+            for(int y=1; y < N; y++){
+                for(int z=1; z < N; z++){
                     if(x < y && y < z &&
                        x * x + y * y == z * z &&
                        std::gcd(std::gcd(x, y), z) == 1){
