+++ optimise/v04b.py
