+++ optimise/v05.py.dis
