+++ optimise/v05_no_inline.py.dis
