+++ optimise/v05_no_inline_func.py.dis
