+++ optimise/v05_no_inline_plus_func.py.dis
