+++ optimise/v08.py
@@ -1,13 +1,11 @@
 import sys
 from math import gcd
-from math import sqrt
 
 
 def process(N):
     combinations = 0
-    max_iter = int(sqrt(N)) + 1
-    for x in range(1, max_iter):
-        for y in range(x + 1, max_iter + 1, 2):
+    for x in range(1, N):
+        for y in range(x + 1, N, 2):
             if gcd(x, y) != 1:
                 continue
             if x * x + y * y > N:  # N -> z
