+++ optimise/v10.py
@@ -7,7 +7,7 @@ def process(N):
     combinations = 0
     max_iter = int(sqrt(N)) + 1
     for x in range(1, max_iter):
-        xxN = int(sqrt(N - x * x))
+        xxN = sqrt(N - x * x)
         for y in range(x + 1, max_iter + 1, 2):
             if gcd(x, y) != 1:
                 continue
