+++ optimise/v11.py
@@ -1,12 +1,6 @@
 import sys
 from math import gcd
 from math import sqrt
-from functools import lru_cache
-
-
-@lru_cache(maxsize=None)
-def GCD(x, y):
-    return gcd(x, y)
 
 
 def process(N):
@@ -15,7 +9,7 @@ def process(N):
     for x in range(1, max_iter):
         xxN = int(sqrt(N - x * x))
         for y in range(x + 1, max_iter + 1, 2):
-            if GCD(x, y) != 1:
+            if gcd(x, y) != 1:
                 continue
             if y > xxN:  # N -> z
                 break
